* No Acc IRQ. Acc Thread returned. No internal ACC kick detection used, MCU checks accelerations every 20ms and decides if kick occured. * Argentum seq updated to decrease amount of false detection situations

+++ QuettaStone_fw/KickList.h
@@ -44,7 +44,7 @@ const uint32_t ksqB[] = {SHRT, LONG, LONG, SHRT, LONG};
 const uint32_t ksqC[] = {SHRT, LONG, LONG, LONG, SHRT, SHRT};
 #elif defined ARGENTUM
 const uint32_t ksqA[] = {LONG, LONG, SHRT};
-const uint32_t ksqB[] = {SHRT, SHRT, LONG, SHRT, SHRT};
+const uint32_t ksqB[] = {SHRT, SHRT, LONG, SHRT, LONG};
 const uint32_t ksqC[] = {SHRT, LONG, SHRT, LONG, SHRT, LONG};
 #endif
